requirements.txt
Flask>=3.0.0
websocket-client>=1.6.0
